# Model/LotListItem.php
<?php
/**
 * LotListItem
 *
 * PHP version 8.1.1
 *
 * @category Class
 * @package  OpenAPI\Server\Model
 * @author   OpenAPI Generator team
 * @link     https://github.com/openapitools/openapi-generator
 */

/**
 * cars
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://github.com/openapitools/openapi-generator.git
 *
 */

/**
 * NOTE: This class is auto generated by the openapi generator program.
 * https://github.com/openapitools/openapi-generator
 * Do not edit the class manually.
 */

namespace OpenAPI\Server\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\Accessor;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the LotListItem model.
 *
 * @package OpenAPI\Server\Model
 * @author  OpenAPI Generator team
 */

class LotListItem 
{
        /**
     * @var int|null
     * @SerializedName("id")
     * @Type("int")
    */
    #[Assert\Type("int")]
    protected ?int $id = null;

    /**
     * @var string|null
     * @SerializedName("manufacturer")
     * @Type("string")
    */
    #[Assert\Type("string")]
    protected ?string $manufacturer = null;

    /**
     * @var string|null
     * @SerializedName("model")
     * @Type("string")
    */
    #[Assert\Type("string")]
    protected ?string $model = null;

    /**
     * @var float|null
     * @SerializedName("price")
     * @Type("float")
    */
    #[Assert\Type("float")]
    protected ?float $price = null;

    /**
     * @var int|null
     * @SerializedName("year")
     * @Type("int")
    */
    #[Assert\Type("int")]
    protected ?int $year = null;

    /**
     * @var float|null
     * @SerializedName("engine_volume")
     * @Type("float")
    */
    #[Assert\Type("float")]
    protected ?float $engineVolume = null;

    /**
     * @var string[]|null
     * @SerializedName("images")
     * @Type("array<string>")
    */
    #[Assert\All([
        new Assert\Type("string"),
    ])]
    protected ?array $images = null;

    /**
     * @var bool|null
     * @SerializedName("is_sold")
     * @Type("bool")
    */
    #[Assert\Type("bool")]
    protected ?bool $isSold = false;

    /**
     * Constructor
     * @param array|null $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        if (is_array($data)) {
            $this->id = array_key_exists('id', $data) ? $data['id'] : $this->id;
            $this->manufacturer = array_key_exists('manufacturer', $data) ? $data['manufacturer'] : $this->manufacturer;
            $this->model = array_key_exists('model', $data) ? $data['model'] : $this->model;
            $this->price = array_key_exists('price', $data) ? $data['price'] : $this->price;
            $this->year = array_key_exists('year', $data) ? $data['year'] : $this->year;
            $this->engineVolume = array_key_exists('engineVolume', $data) ? $data['engineVolume'] : $this->engineVolume;
            $this->images = array_key_exists('images', $data) ? $data['images'] : $this->images;
            $this->isSold = array_key_exists('isSold', $data) ? $data['isSold'] : $this->isSold;
        }
    }

    /**
     * Gets id.
     *
     * @return int|null
     */
    public function getId(): ?int
    {
        return $this->id;
    }

    /**
    * Sets id.
    *
    * @param int|null $id
    *
    * @return $this
    */
    public function setId(?int $id = null): self
    {
        $this->id = $id;

        return $this;
    }




    /**
     * Gets manufacturer.
     *
     * @return string|null
     */
    public function getManufacturer(): ?string
    {
        return $this->manufacturer;
    }

    /**
    * Sets manufacturer.
    *
    * @param string|null $manufacturer
    *
    * @return $this
    */
    public function setManufacturer(?string $manufacturer = null): self
    {
        $this->manufacturer = $manufacturer;

        return $this;
    }




    /**
     * Gets model.
     *
     * @return string|null
     */
    public function getModel(): ?string
    {
        return $this->model;
    }

    /**
    * Sets model.
    *
    * @param string|null $model
    *
    * @return $this
    */
    public function setModel(?string $model = null): self
    {
        $this->model = $model;

        return $this;
    }




    /**
     * Gets price.
     *
     * @return float|null
     */
    public function getPrice(): ?float
    {
        return $this->price;
    }

    /**
    * Sets price.
    *
    * @param float|null $price
    *
    * @return $this
    */
    public function setPrice(?float $price = null): self
    {
        $this->price = $price;

        return $this;
    }




    /**
     * Gets year.
     *
     * @return int|null
     */
    public function getYear(): ?int
    {
        return $this->year;
    }

    /**
    * Sets year.
    *
    * @param int|null $year
    *
    * @return $this
    */
    public function setYear(?int $year = null): self
    {
        $this->year = $year;

        return $this;
    }




    /**
     * Gets engineVolume.
     *
     * @return float|null
     */
    public function getEngineVolume(): ?float
    {
        return $this->engineVolume;
    }

    /**
    * Sets engineVolume.
    *
    * @param float|null $engineVolume
    *
    * @return $this
    */
    public function setEngineVolume(?float $engineVolume = null): self
    {
        $this->engineVolume = $engineVolume;

        return $this;
    }




    /**
     * Gets images.
     *
     * @return string[]|null
     */
    public function getImages(): ?array
    {
        return $this->images;
    }

    /**
    * Sets images.
    *
    * @param string[]|null $images
    *
    * @return $this
    */
    public function setImages(?array $images = null): self
    {
        $this->images = $images;

        return $this;
    }




    /**
     * Gets isSold.
     *
     * @return bool|null
     */
    public function isIsSold(): ?bool
    {
        return $this->isSold;
    }

    /**
    * Sets isSold.
    *
    * @param bool|null $isSold
    *
    * @return $this
    */
    public function setIsSold(?bool $isSold = null): self
    {
        $this->isSold = $isSold;

        return $this;
    }



}