# openapi.yaml
openapi: 3.0.3
info:
  title: cars
  version: 1.0.0
paths:
...
  /catalog:
    get:
      operationId: getCatalog
      tags: [Catalog]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: search
          in: query
          schema: { type: string }
        - name: manufacturer_id
          in: query
          schema: { type: integer }
        - name: model_id
          in: query
          schema: { type: integer }
        - name: color_id
          in: query
          schema: { type: integer }
        - name: transmission
          in: query
          schema: { type: string, enum: ["Автомат", "Механика", "Робот", "Вариатор"] }
        - name: drive
          in: query
          schema: { type: string, enum: ["Полный", "Передний", "Задний"] }
        - name: year
          in: query
          schema: { type: integer }
        - name: price_from
          in: query
          schema: { type: number }
        - name: price_to
          in: query
          schema: { type: number }
        - name: mileage_from
          in: query
          schema: { type: integer }
        - name: mileage_to
          in: query
          schema: { type: integer }
        - name: engine_volume_id
          in: query
          schema: { type: integer }
        - name: is_sold
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: список машин
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LotListItem'
    post:
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LotDetail' }
      responses:
        '201':
          description: машина добавлена
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LotListItem:
      type: object
      properties:
        id: { type: integer }
        manufacturer: { type: string }
        model: { type: string }
        price: { type: number }
        year: { type: integer }
        engine_volume: { type: number }
        images: { type: array, items: { type: string } }
        is_sold: { type: boolean, default: false }
    LotDetail:
      type: object
      properties:
        id: { type: integer }
        manufacturer: { type: string }
        model: { type: string }
        year: { type: integer }
        price: { type: number }
        mileage: { type: integer }
        engine_volume: { type: number }
        color: { type: string }
        transmission: { type: string }
        drive: { type: string }
        body_number: { type: string }
        is_sold: { type: boolean, default: false }
        sold_date: { type: string, format: date, nullable: true }
        review:
          oneOf:
            - $ref: '#/components/schemas/Review'
          nullable: true
        images: { type: array, items: { type: string } }
...
    Review:
      type: object
      properties:
        id: { type: integer }
        lot_id: { type: integer }
        manufacturer: { type: string }
        model: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        avatar_url: { type: string, nullable: true }
        rating: { type: integer }
        comment: { type: string }
        created_at: { type: string, format: date-time }